Добавлена возможность работы с заказами

+++ UltraLogger.Core/Application/DTOs/OrderDTO.cs
@@ -6,7 +6,7 @@ public class OrderDTO
     public string Number { get; set; } = string.Empty;
     public long CustomerId { get; set; }
     public string SteelGrade { get; set; } = string.Empty;
-    public int PlateThickness { get; set; }
+    public float PlateThickness { get; set; }
     public int PlateWidth { get; set; }
     public int PlateLength { get; set; }
 }
